chore: move node definitions
