+++ src/shaderityShaders/nodeDefinitions/Artistic/Normal/normalTextureScale.glsl
