+++ src/shaderityShaders/nodeDefinitions/Artistic/Occlusion/occlusionStrength.glsl
