+++ src/shaderityShaders/nodeDefinitions/Channel/AddElements/floatToVector4.glsl
