+++ src/shaderityShaders/nodeDefinitions/Channel/Combine/combineVector3.glsl
