+++ src/shaderityShaders/nodeDefinitions/Channel/Combine/combineVector4.glsl
