+++ src/shaderityShaders/nodeDefinitions/Channel/RemoveElements/mat3ToMat2.glsl
