+++ src/shaderityShaders/nodeDefinitions/Channel/RemoveElements/mat4ToMat2.glsl
