+++ src/shaderityShaders/nodeDefinitions/Channel/RemoveElements/mat4ToMat3.glsl
