+++ src/shaderityShaders/nodeDefinitions/Channel/RemoveElements/vector3ToVector2.glsl
