+++ src/shaderityShaders/nodeDefinitions/Channel/RemoveElements/vector4ToVector2.glsl
