+++ src/shaderityShaders/nodeDefinitions/Channel/RemoveElements/vector4ToVector3.glsl
