+++ src/shaderityShaders/nodeDefinitions/Channel/Split/splitVector3.glsl
