+++ src/shaderityShaders/nodeDefinitions/Channel/Split/splitVector4.glsl
