+++ src/shaderityShaders/nodeDefinitions/Input/Basic/boolean.glsl
