+++ src/shaderityShaders/nodeDefinitions/Input/Basic/uniformVector3.glsl
