+++ src/shaderityShaders/nodeDefinitions/Input/Basic/uniformVector4.glsl
