+++ src/shaderityShaders/nodeDefinitions/Input/Geometry/inputPosition.vs.glsl
