+++ src/shaderityShaders/nodeDefinitions/Input/Texture/sampleTexture2D.glsl
