+++ src/shaderityShaders/nodeDefinitions/Utility/Output/fragmentColor.fs.glsl
